Enum poke_a_mango::Error

pub enum Error {
    FileParsingFailed {
        desc: &'static str,
        errors: Vec<String>,
    },
    Io {
        desc: &'static str,
        op: &'static str,
    },
}

Enum representing all possible ways the application can fail.

Variants
FileParsingFailed

Failed to parse the specified file because of the specified errors.

Fields

desc: &'static str

The file that failed to parse.

errors: Vec<String>

The parsing errors that occured.

Io

An I/O error occured.

This includes higher-level I/O errors like FS ones.

Fields

desc: &'static str

The file that failed to parse.

op: &'static str

The failed operation.

This should be lowercase and imperative ("create", "open").

Methods

impl Error

fn print_error<W: Write>(&self, err_out: &mut W)

Get the executable exit value from an Error instance.

Examples

let mut out = Vec::new();
Error::Io {
    desc: "leaderboard",
    op: "write",
}.print_error(&mut out);
assert_eq!(String::from_iter(out.iter().map(|&i| i as char)),
           "Writing leaderboard failed.\n".to_string());

fn exit_value(&self) -> i32

Get the executable exit value from an Error instance.

Examples

assert_eq!(Error::FileParsingFailed {
    desc: "",
    errors: vec![],
}.exit_value(), 1);