Module poke_a_mango::ops

Main functions doing actual work.

Use create_window() to set up a window of correct size, then set up a conrod::Ui on it like you'd do normally. Then create an instance of Widgets and call update() on it each update event, then check for states that need manual handling from usercode and act accordingly.

Modules

state
Manual GameState manipulation, to be used when implementing a UI.

Structs

Leader
A leaderboard entry.

Widgets
Container for all widgets' IDs, also manages setting them.

Enums

Difficulty
The game's difficulty, chosen in the ChooseDifficulty step.

GameState
Game's all possible states.

Functions

create_window
Create a window based on the size of the target desktop.

set_button_style
Set the button style to a one that looks :+1: on the canvas.

window_size
Get the window size from the size of the target desktop.