Module poke_a_mango::ops::state

Manual GameState manipulation, to be used when implementing a UI.

For a UI implementing the use of this see the implementation of Widgets::update().

Functions

end_mango
End the mango processing sequence, given whether it was clicked and whether it was actually a mango.

press_back
Press the Back button in ChooseDifficulty, GameOver or DisplayLeaderboard, transforming them into MainMenu.

press_display_highscores
Press the Display Highscores button in MainMenu, transforming it into LoadLeaderboard.

press_exit
Press the Exit button in MainMenu, transforming it into Exit.

press_start
Press the Start button in MainMenu, transforming it into ChooseDifficulty.

select_difficulty
Press one of the Difficulty buttons in ChooseDifficulty, transforming it into Exit.

submit_name

tick_mango
Tick the mango in Playing state.