Struct poke_a_mango::ops::Leader

pub struct Leader {
    pub name: String,
    pub time: DateTime<FixedOffset>,
    pub score: u64,
}

A leaderboard entry.

Examples

Reading a leaderboard, adding an entry to it, then writing it back.

let tf = temp_dir().join("poke-a-mango-doctest").join("ops-Leader-0");
create_dir_all(&tf).unwrap();

let tf = tf.join("leaderboard.toml");
File::create(&tf).unwrap();

let mut leaders = Leader::read(&tf).unwrap();
assert!(leaders.is_empty());
leaders.push(Leader::now("nabijaczleweli".to_string(), 105));
assert_eq!(Leader::write(leaders, &tf), Ok(()));

This could alternatively be done with Leader::append().

Fields
name: String

Name the user entered when prompted.

time: DateTime<FixedOffset>

Time the score was achieved.

score: u64

User's score.

Methods

impl Leader

fn now(name: String, score: u64) -> Leader

Create a new leaderboard entry with time set to now.

Examples

let ldr = Leader::now("nabijaczleweli".to_string(), 36);
assert_eq!(ldr.name, "nabijaczleweli");
assert_eq!(ldr.score, 36);

fn read(p: &Path) -> Result<Vec<Leader>, Error>

Read leaderboard from the specified file.

If the specified file doesn't exist an empty leaderboard is returned.

Examples

let tf = temp_dir().join("poke-a-mango-doctest").join("ops-Leader-read-0");
create_dir_all(&tf).unwrap();

let tf = tf.join("leaderboard.toml");
File::create(&tf).unwrap();

assert_eq!(Leader::read(&tf), Ok(vec![]));

fn write(queued_leaders: Vec<Leader>, p: &Path) -> Result<(), Error>

Save leaderboard to the specified file.

Examples

let tf = temp_dir().join("poke-a-mango-doctest").join("ops-Leader-write-0");
create_dir_all(&tf).unwrap();

Leader::write(vec![Leader::now("nabijaczleweli".to_string(), 105),
                   Leader::now("skorezore".to_string(), 51)],
              &tf.join("leaderboard.toml"))
    .unwrap();

fn append(ldr: Leader, p: &Path) -> Result<(), Error>

Append the specified leader to the leaderboard file at the specified path.

This is equivalent to reading it, appending the leader, properly sorting and writing it back.

Examples

let tf = temp_dir().join("poke-a-mango-doctest").join("ops-Leader-append-0");
create_dir_all(&tf).unwrap();

assert_eq!(Leader::append(Leader::now("nabijaczleweli".to_string(), 105), &tf.join("leaderboard.toml")), Ok(()));